requirements.txt
kafka-python==2.0.2
pandas==2.1.4
streamlit==1.29.0
pydantic==2.5.0
#numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0



